[Git page]Step 2: Added the remaining git functions in a separate file and imported them. Status: Everything works!
